feat: Added Point and Entity Class
