+++ src/App.java
@@ -1,3 +1,5 @@
 public class App {
-    
+    public static void main(String[] args) {
+        System.out.println("Hello, World!");
+    }
 }
