+++ src/models/Tile.java
@@ -1,3 +1,7 @@
+package Models;
+
+import Utils.TileTypes;
+
 public class Tile {
     private TileTypes type;
