+++ src/models/Transport.java
@@ -1,3 +1,7 @@
+package Models;
+
+import Utils.TransportTypes;
+
 public class Transport {
     private TransportTypes type;
